Add simple docker-compose development environment
Steps to run dev environment:

1. docker compose build
2. docker compose up

The app container will update the poetry virtualenv on each start and migrate the database. This is
done by the docker/entrypoint-dev.sh. It uses a separate virtualenv in the container so one can
develop on the host machine with IDEs as usual without conflicts in the container.

This means that changing something related to dependencies, the container needs to be rebuild or
restarted to reflect the changes.

Future improvements:

- make Django 12factor compatible (as in configure via ENV)
- use PostgreSQL instead of SQLite and get rid of the .scratch dir
- integrate Renovate to keep Dockerfile and friends up-to-date